requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
pymongo==4.6.1
motor==3.3.2
instaloader==4.10.3
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
dnspython==2.4.2
